backend/requirements.txt
Django==3.2.8
django-seed==0.3.1
djangorestframework==3.12.4
mysqlclient==2.0.3
bs4==0.0.1
requests==2.26.0
pandas==1.3.3
openpyxl==3.0.9